pipeline-code/llm-worker/llm_main.py
import os
import json
import yaml

NLP_OUTPUT_FILE = "/data/nlp-output.json"
LLM_REPORT_FILE = "/data/llm-report.json"

ENV_DOC_PATH     = "/data/docs/env_overview.yaml"
PDF_REFERENCES   = "/data/docs/arch_diagram.pdf"   # Example placeholders
IMAGE_REFERENCES = "/data/docs/arch_diagram.jpg"   # You'd need external libs to parse images

def load_environment_overview():
    """
    Loads the environment overview from a YAML file mounted via ConfigMap.
    If you also want to parse PDF or images, you'd integrate OCR or pdf2text here.
    For now, we just read the YAML text for LLM context.
    """
    if os.path.exists(ENV_DOC_PATH):
        with open(ENV_DOC_PATH, "r") as f:
            try:
                return yaml.safe_load(f)
            except yaml.YAMLError:
                print("Warning: Could not parse env_overview.yaml")
                return {}
    else:
        print("No environment overview file found.")
        return {}

def run_multi_agent_rca(nlp_data, env_overview):
    """
    Example function that merges the NLP output with environment data.
    In reality, you'd pass these to the LLM with your agent role prompts.
    """
    combined_context = {
        "nlp_summary": nlp_data,
        "environment_details": env_overview
    }

    rca_result = {
        "root_cause": "Network misconfiguration or ephemeral resource issue (example)",
        "recommendation": "Check the environment doc for load balancer config, ensure node autoscaler is scaled properly."
    }

    return {
        "combined_context": combined_context,
        "rca_analysis": rca_result
    }

def main():
    # 1) Check if NLP output file exists
    if not os.path.exists(NLP_OUTPUT_FILE):
        print("No NLP output found; skipping LLM stage.")
        return

    # 2) Load NLP output (e.g., summarized logs)
    with open(NLP_OUTPUT_FILE, "r") as f:
        nlp_data = json.load(f)

    # 3) Load environment overview (includes references to PDFs, images, etc.)
    env_overview = load_environment_overview()

    # 4) Run multi-agent or single-agent logic, passing environment doc for context
    final_data = run_multi_agent_rca(nlp_data, env_overview)

    # 5) Write final LLM report
    with open(LLM_REPORT_FILE, "w") as out:
        json.dump(final_data, out, indent=2)

    print("LLM stage completed. Final report saved at:", LLM_REPORT_FILE)

if __name__ == "__main__":
    main()
