pipeline-code/llm-worker/requirements.txt
torch==1.13.1
transformers==4.26.1
pyyaml==6.0
